requirements.txt
requests>=2.28.1
python-logstash>=0.4.8

